Adding back collider's workaround so level prefab saves the most current data (#10231)
Adding workaround back now that this has been fixed https://github.com/o3de/o3de/pull/10222

The workaround fixes a bug where physx collider material list was not saved correctly in a prefab level. Marking entity as dirty fixes the issue.

There is a side effect to this fix though, the undo stack needs to be amended and there is no good way to do that at the moment. This means a user will have to hit Ctrl+Z twice to revert its last change, which is not good, but not as bad as losing data. Another GHI will be created to improve on this.

Fixes #9780 

Tests done:
- Followed the steps from GHI and checked the prefab level saves the most current data from the colliders
- Followed the steps from GHI's comment and checked all entities remain under "level" node in entity outliner after creating a prefab with a collider in it.
- Run prefabs automated tests 5 times locally. All passed.

Signed-off-by: moraaar <moraaar@amazon.com>