Parameterize ToolsApplicationFixture to allow it to not check for leaks on destruction
The PhysX tests create a ToolsApplication in its environment, which trips
up the leak detection functionality. By making it a template parameter of
the fixture, the PhysX tests can instruct the fixture to not check the
allocators for leaks at the end of each individual test. The global
environment still checks for leaks after the tests are finished.

Signed-off-by: Chris Burel <burelc@amazon.com>