Add AZ_CLASS_ALLOCATOR macros to subclasses as appropriate
The AZ_CLASS_ALLOCATOR macro hard-codes the alignment used in the call
to the underlying allocator's allocate method to the alignment of Base:
```
return AZ::AllocatorInstance< _Allocator >::Get().allocate(size, AZStd::alignment_of< _Class >::value);
```

If a derived class increases the alignment requirement,
`Base::operator new(size_t)` has no idea, and supplies the alignment of
Base to the allocate function. But only up to a point: if the alignment
of Derived is greater than `__STDCPP_DEFAULT_NEW_ALIGNMENT__` (which is
16), then the operator new overload that takes both a size and alignment
come into play. At that point, the correct alignment is provided to the
function as a parameter, and everything works again.

Attempting to store a misaligned address to a type that requires a
higher alignment may crash. But that only manifests itself when
alignof(Base) is 8 and alignof(Derived) is 16.

This would be causing problems all over the place, if it weren't for the
Serialize::InstanceFactory. The InstanceFactory ignores inherited
operator new() methods and checks if a type T directly has used the
AZ_CLASS_ALLOCATOR macro. If it has, it uses new to create the instance
(aznew is now synonymous with new). If it didn't, it uses the real
alignment requirement of T, but forces the instance to be allocated with
the SystemAllocator. This gets around the incorrect alignment hard-coded
in Base::operator new(size), but if Base is configured to use a
different allocator, it would cause its own set of problems when delete
derived; is called.

I found this in action in the ConvertEmissiveUnitFunctorSourceData
class. Its AZ::Vector2 instance increases its alignment requirement to
16, when the base class MaterialFunctorSourceData's alignment is 8. This
means that simply writing new ConvertEmissiveUnitFunctorSourceData() may
crash. In practice though, nothing does this, because all the instances
are made by the Serialize::InstanceFactory.

This fixes the above problem by adding the AZ_CLASS_ALLOCATOR macro to
subclasses that don't have it. These were identified with the following
matcher to clang-query:

```
let hasOverloadedNew hasMethod(cxxMethodDecl(hasName("operator new"), parameterCountIs(1), anyOf(isExpandedFromMacro("AZ_CLASS_ALLOCATOR"), isExpandedFromMacro("AZ_CLASS_ALLOCATOR_DECL"))))
match cxxRecordDecl(isDefinition(), unless(hasOverloadedNew), has(fieldDecl()), hasAnyBase(hasType(cxxRecordDecl( hasOverloadedNew ))))
```

Signed-off-by: Chris Burel <burelc@amazon.com>