Add box dimensions setter/getter to EditorPrimitiveColliderComponentRequests (#14983)
Signed-off-by: moraaar <moraaar@amazon.com>

## What does this PR do?

`EditorPrimitiveColliderComponentRequests` had functions to set the parameters of all the shapes except for box. Before code was using `BoxManipulatorRequestBus` to get/set the box dimensions, which is more appropriate to handle as manipulator rather to set the shape dimentions.

## How was this PR tested?

Changed the test code of Editor Primitive Collider component to use the new functions in `EditorPrimitiveColliderComponentRequests` bus.
Run the Editor PhysX tests and it passes.